Base Look and feel - Heavy commit
While the development meets the functionality, and the time-pickers have a good UX/UI, the base component of the project was horrible. For this reason the whole base of the project (appComponent) was restructured prioritizing a better UI. New custom components will also be added as the need arises. From this commit the DEVELOP branch will be created.